refactor: 시간 형식 맞춰주는 함수 분리
