+++ src/pages/statistics/all/AllStatisticsPage.tsx
@@ -12,6 +12,11 @@ import dayjs from 'dayjs';
 import * as S from '@/pages/statistics/all/AllStatisticsPage.style';
 import { fetchAllAverage, fetchAllGraph } from '@/apis/statistics.api';
 import { IoIosArrowBack, IoIosArrowForward } from 'react-icons/io';
+import {
+  formatAverageTime,
+  formatHours,
+  formatHoursAndMinutes,
+} from '@/utils/formatTime';
 
 const CustomLegend = () => (
   <S.LegendContainer>
@@ -51,23 +56,6 @@ export default function AllStatisticsPage() {
       lastMonth: { hours: '', minutes: '' },
     },
   });
-
-  const formatAverageTime = (hours: string, minutes: string) => {
-    const formattedHours = parseInt(hours, 10);
-    const formattedMinutes = parseInt(minutes, 10);
-    return `${formattedHours}시간 ${formattedMinutes}분`;
-  };
-
-  const formatHoursAndMinutes = (totalMinutes: number) => {
-    const hours = Math.floor(totalMinutes / 60);
-    const minutes = totalMinutes % 60;
-    return `${hours}시간 ${minutes}분`;
-  };
-
-  const formatHours = (totalMinutes: number): string => {
-    const hours = Math.floor(totalMinutes / 60);
-    return `${hours}시간`;
-  };
 
   const loadGraphData = async () => {
     try {
